Add grammar for shorthand callable type expressions
Rerun `make regen-pgen` as needed.

This grammar is smallar than what's proposed in some of the PEP
docs because it doesn't include PEP 646 support. This is intentional,
because PEP 646 is still under consideration and the details of how
to handle the AST for unpack operations are not yet decided.

It does include support for PEP 612 param specs though.

There are not yet tests because we still need a stubbed out
validation pass in order to actually produce the AST, I'll do
that in the next commit.