Add a temporary module for more-readable tests of operator binding
Summary:

We won't merge this module, all of the cases are already covered
by `test_ast.py`. But it's not easy to read the blobs of expected
parse results in `test_ast.py`. This test, which compares parsed
results with and without explicit parentheses, will probably be
much easier to skim, and I am planning to parachute the examples
verbatim into the PEP.

It's a little hacky to do a string comparison on ast.dump but that's
the quickest way to verify this (the actual ast data includes locations
and is therefore not equal). Since we'll eliminate this module before
any final merge it shouldn't matter.

Test Plan:

```
> ./python Lib/test/test_callable_type_syntax_for_pep.py
.
----------------------------------------------------------------------
Ran 1 test in 0.002s

OK
```