Forbid `...` as an argument type expression
Summary:

As a rule, `...` is a valid expression, so by default it would be
permitted as an argument type.

But it is never a semantically valid type, so even though
the syntax would normally allow it there's no particular reason
why the syntax should permit it.

And in the context of callable type annotations there are *very* good
reasons why we benefit from disallowing it, because it's overloaded
to mean other things:
- The expression `(...) -> bool` does *not* indicate an arguments
  list with a type of `...`, it means AnyArguments which is special
- Also in function signatures (which are closely tied to typing because
  of both stubs and Callback Protocols) the value `...` often is used
  as a placeholder for a default
- We'd prefer not to allow `(...,) -> bool` if possible... that's actually
  why we noticed this issue, we were testing edge cases involving trailing
  commas and the earlier grammar would match `(...,) -> bool`, but as
  an arguments list rather than an AnyArguments.

Test Plan:

```
./python Lib/test/test_ast.py
```